Changes to config directory specification
This is now a function, which allows it to be changed more easily during
tests, and the search path now works slightly differently. In the CLI
tools like `update`, the PWD is searched, but when running the app
normally, the default config won't search the PWD.

The exact search behavior can be overriden with `AF_CONFIG_DIR`, which
specifies the single directory where the config actually lives.