Add media rendering into the front-end
This adds the media renderer triggered by requests to the RSS feeds.
There's a bunch of refactoring going on here because now we have two
slightly different ways of generating RSS feeds — one for when we're
just looking at the files on disk, and one for when we're predicting
what files will be available when the renderer finishes its work.