Adjust threading and add logging
To prevent deadlocks, and because we get most of the bang for our buck
when throwing the ffmpeg-related stuff to threads, this changes around
what gets done in new threads.

- New entries are created in the main thread, but their hash
  calculations and metadata updating are passed to the executor.
- Checking existing entries for file metadata (and updating the file
  metadata if it doesn't exist) is given its own thread, but the
  updating entry metadata is *not* done in that thread.
- Entries with updated file metadata are collected in the main thread,
  but the metadata and hashes are updated in new threads.

When the updater is first run on a large number of files, we'll have a
huge amount of work to do, and how we split it up probably won't make a
*major* difference so long as most of the ffmpeg calls are in background
threads. In the steady state, most of the work will likely be in
verifying that `entry.files` is populated, and verifying that all the
hashes are still valid. We may want to move the "check if `entry.files` is
populated" check into the main thread if we're finding that waiting on
the GIL is adding too much overhead there.

It's a bit hard to debug the threading-related issues, so I'm adding
some debugging log messages here as well.