Fix duration issue and add test for weird names
The test for handling apostrophes turned up another bug in
_merge_subsets stemming from the fact that we modify the FileInfo for
segmentable files to make it easier to merge them later.

Clearly this is a typing issue where FileInfo is overloaded in one way
or another and it's hard to keep track of how we are using it, but for
now we'll just keep fixing the bugs as they come up.