Delete objects to be replaced
SQLAlchemy maps specific instances to the objects in the database. If
the object is mutated in-place, adding it to the table again will issue
an UPDATE, but if the object is replaced with a new instance, you need
to delete it first, then re-add it, otherwise it will fail the UNIQUE
constraint, since it's trying to add the same thing twice.